Add linux and mingw x64 target to the paging extension (#4280)
* Add linux and mingw x64 target to the paging extension

**Note:**
The version of `paging-mpp` still needs updating once https://github.com/cashapp/multiplatform-paging/pull/121 is in a release

* Use release version of paging-mpp

* Replace sqlite static lib extraction with cinterop of the sqlite amalgamation

---------
