+++ extensions/androidx-paging3/build.gradle
@@ -1,3 +1,6 @@
+import org.jetbrains.kotlin.konan.target.HostManager
+import org.jetbrains.kotlin.gradle.plugin.mpp.NativeOutputKind
+
 plugins {
   alias(libs.plugins.kotlin.multiplatform)
   alias(libs.plugins.publish)
@@ -14,6 +17,8 @@ kotlin {
   js {
     browser()
   }
+  linuxX64()
+  mingwX64()
 
   sourceSets {
     commonMain {
@@ -30,11 +35,21 @@ kotlin {
         implementation libs.stately.concurrency
       }
     }
-    iosMain {
+    nativeMain {
     }
-    iosTest {
+    nativeTest {
+      dependsOn(nativeMain)
+      dependsOn(commonTest)
       dependencies {
         implementation projects.drivers.nativeDriver
+      }
+    }
+    iosMain {
+      dependsOn(nativeMain)
+    }
+    iosTest {
+      dependsOn(nativeTest)
+      dependencies {
         implementation libs.app.cash.paging.runtime.uikit
       }
     }
@@ -43,6 +58,18 @@ kotlin {
     }
     iosSimulatorArm64Test {
       dependsOn(iosTest)
+    }
+    linuxX64Main {
+      dependsOn(nativeMain)
+    }
+    linuxX64Test {
+      dependsOn(nativeTest)
+    }
+    mingwX64Main {
+      dependsOn(nativeMain)
+    }
+    mingwX64Test {
+      dependsOn(nativeTest)
     }
     jvmTest {
       dependencies {
@@ -59,9 +86,43 @@ kotlin {
 
   configure([targets.iosX64, targets.iosArm64, targets.iosSimulatorArm64]) {
     binaries.configureEach {
-      linkerOpts += ["-lsqlite3"]
+      // we only need to link sqlite for the test binaries
+      if (outputKind == NativeOutputKind.TEST) {
+        linkerOpts += ["-lsqlite3"]
+      }
+    }
+  }
+
+  configure([targets.linuxX64]) {
+    compilations.configureEach {
+      if (name == "test") {
+        cinterops {
+          sqlite {
+            // use sqlite3 amalgamation on linux tests to prevent linking issues on new linux distros with dependency libraries which are to recent (for example glibc)
+            // see: https://github.com/touchlab/SQLiter/pull/38#issuecomment-867171789
+            defFile = new File("$rootDir/libs/linux/cinterop/sqlite3.def")
+          }
+        }
+      }
+    }
+  }
+
+  configure([targets.mingwX64]) {
+    binaries.configureEach {
+      // we only need to link sqlite for the test binaries
+      if (outputKind == NativeOutputKind.TEST) {
+        linkerOpts += ["-Lc:\\msys64\\mingw64\\lib", "-L$rootDir\\libs\\windows".toString(), "-lsqlite3"]
+      }
     }
   }
 }
+
+//linking fails for the linux test build if not built on a linux host
+//ensure the tests and linking for them is only done on linux hosts
+tasks.named("linuxX64Test") { enabled = HostManager.hostIsLinux }
+tasks.named("linkDebugTestLinuxX64") { enabled = HostManager.hostIsLinux }
+
+tasks.named("mingwX64Test") { enabled = HostManager.hostIsMingw }
+tasks.named("linkDebugTestMingwX64") { enabled = HostManager.hostIsMingw }
 
 apply from: "$rootDir/gradle/gradle-mvn-push.gradle"
