+++ extensions/androidx-paging3/src/nativeTest/kotlin/app/cash/sqldelight/paging3/ProvideDbDriver.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2016 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.sqldelight.paging3
+
+import app.cash.sqldelight.db.AfterVersion
+import app.cash.sqldelight.db.QueryResult
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.db.SqlSchema
+import app.cash.sqldelight.driver.native.inMemoryDriver
+
+private fun defaultSchema(): SqlSchema<QueryResult.Value<Unit>> {
+  return object : SqlSchema<QueryResult.Value<Unit>> {
+    override val version: Int = 1
+    override fun create(driver: SqlDriver) = QueryResult.Unit
+    override fun migrate(
+      driver: SqlDriver,
+      oldVersion: Int,
+      newVersion: Int,
+      vararg callbacks: AfterVersion,
+    ) = QueryResult.Unit
+  }
+}
+
+actual suspend fun provideDbDriver(): SqlDriver =
+  inMemoryDriver(defaultSchema())
