+++ gradle.properties
@@ -23,3 +23,6 @@ kotlin.js.compiler=ir
 kotlin.mpp.stability.nowarn=true
 kotlin.native.ignoreDisabledTargets=true
 org.gradle.jvmargs=-Xmx2g -XX:MaxMetaspaceSize=1g -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
+
+# caches break the linkage of the sqlite amalgamation
+kotlin.native.cacheKind.linuxX64=none
